+++ src/Resources/contao/config/config.php
@@ -25,11 +25,12 @@
 
 $GLOBALS['BE_MOD']['content']['internetschach'] = array
 (
-	'tables'          => array('tl_internetschach', 'tl_internetschach_spieler', 'tl_internetschach_anmeldungen', 'tl_internetschach_tabellen', 'tl_internetschach_preise'),
-	'importCSV'       => array('\Schachbulle\ContaoInternetschachBundle\Classes\Import', 'importCSV'),
-	'importTable'     => array('\Schachbulle\ContaoInternetschachBundle\Classes\Import', 'importTable'),
-	'exportXLS'       => array('\Schachbulle\ContaoInternetschachBundle\Classes\Export', 'getExcel'),
-	'qualifikationen' => array('\Schachbulle\ContaoInternetschachBundle\Classes\Qualifikationen', 'Aktualisieren') 
+	'tables'               => array('tl_internetschach', 'tl_internetschach_spieler', 'tl_internetschach_anmeldungen', 'tl_internetschach_tabellen', 'tl_internetschach_preise'),
+	'importCSV'            => array('\Schachbulle\ContaoInternetschachBundle\Classes\Import', 'importCSV'),
+	'importTable'          => array('\Schachbulle\ContaoInternetschachBundle\Classes\Import', 'importTable'),
+	'exportXLS'            => array('\Schachbulle\ContaoInternetschachBundle\Classes\Export', 'getExcel'),
+	'qualifikationen'      => array('\Schachbulle\ContaoInternetschachBundle\Classes\Qualifikationen', 'Aktualisieren'),
+	'finalqualifikationen' => array('\Schachbulle\ContaoInternetschachBundle\Classes\Qualifikationen', 'FinaleAktualisieren')
 );
 
 
@@ -42,7 +43,7 @@ $GLOBALS['TL_CTE']['includes']['internetschach_formular'] = 'Schachbulle\ContaoI
 $GLOBALS['TL_CTE']['includes']['internetschach_anmeldungen'] = 'Schachbulle\ContaoInternetschachBundle\ContentElements\Anmeldungen';
 $GLOBALS['TL_CTE']['includes']['internetschach_tabelle'] = 'Schachbulle\ContaoInternetschachBundle\ContentElements\Tabelle';
 
-if(TL_MODE == 'BE') 
+if(TL_MODE == 'BE')
 {
-	$GLOBALS['TL_CSS'][] = 'bundles/contaointernetschach/css/be.css'; 
-} 
+	$GLOBALS['TL_CSS'][] = 'bundles/contaointernetschach/css/be.css';
+}