+++ src/Resources/contao/dca/tl_internetschach_anmeldungen.php
@@ -53,6 +53,13 @@ $GLOBALS['TL_DCA']['tl_internetschach_anmeldungen'] = array
 		),
 		'global_operations' => array
 		(
+			'finalqualifikationen' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_internetschach_tabellen']['finalqualifikationen'],
+				'href'                => 'key=finalqualifikationen',
+				'icon'                => 'bundles/contaointernetschach/images/qualifikationen.png',
+				'attributes'          => 'onclick="if(!confirm(\'' . $GLOBALS['TL_LANG']['tl_internetschach_anmeldungen']['finalqualifikationen_confirm'] . '\'))return false;Backend.getScrollOffset()"',
+			),
 			'all' => array
 			(
 				'label'               => &$GLOBALS['TL_LANG']['MSC']['all'],